main/requirements_wavelet.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
# created-by: conda 25.3.1
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_gnu
alsa-lib=1.2.14=hb9d3cd8_0
aom=3.6.1=h59595ed_0
blas=1.0=mkl
brotli=1.1.0=hb9d3cd8_2
brotli-bin=1.1.0=hb9d3cd8_2
brotli-python=1.1.0=py39hf88036b_2
bzip2=1.0.8=h4bc722e_7
ca-certificates=2025.7.9=hbd8a1cb_0
cairo=1.18.4=h3394656_0
certifi=2025.6.15=pyhd8ed1ab_0
cffi=1.17.1=py39h15c3d72_0
charset-normalizer=3.4.2=pyhd8ed1ab_0
colorama=0.4.6=pyhd8ed1ab_1
contourpy=1.3.0=py39h74842e3_2
cpython=3.9.22=py39hd8ed1ab_1
cuda-cudart=11.8.89=0
cuda-cupti=11.8.87=0
cuda-libraries=11.8.0=0
cuda-nvrtc=11.8.89=0
cuda-nvtx=11.8.86=0
cuda-opencl=12.9.19=0
cuda-runtime=11.8.0=0
cuda-version=12.9=3
cycler=0.12.1=pyhd8ed1ab_1
cyrus-sasl=2.1.27=h54b06d7_7
dbus=1.16.2=h3c4dab8_0
double-conversion=3.3.1=h5888daf_0
einops=0.8.1=pyhd8ed1ab_0
ffmpeg=4.4.2=gpl_hdf48244_113
filelock=3.18.0=pyhd8ed1ab_0
font-ttf-dejavu-sans-mono=2.37=hab24e00_0
font-ttf-inconsolata=3.000=h77eed37_0
font-ttf-source-code-pro=2.038=h77eed37_0
font-ttf-ubuntu=0.83=h77eed37_3
fontconfig=2.15.0=h7e30c49_1
fonts-conda-ecosystem=1=0
fonts-conda-forge=1=0
fonttools=4.58.1=py39h9399b63_0
freetype=2.13.3=ha770c72_1
fsspec=2025.3.2=pyhd8ed1ab_0
giflib=5.2.2=hd590300_0
gmp=6.3.0=hac33072_2
gmpy2=2.2.1=py39h7196dd7_0
gnutls=3.7.9=hb077bed_0
graphite2=1.3.13=h59595ed_1003
h2=4.2.0=pyhd8ed1ab_0
harfbuzz=11.2.1=h3beb420_0
hpack=4.1.0=pyhd8ed1ab_0
huggingface_hub=0.31.4=pyhd8ed1ab_0
hyperframe=6.1.0=pyhd8ed1ab_0
icu=75.1=he02047a_0
idna=3.10=pyhd8ed1ab_1
importlib-resources=6.5.2=pyhd8ed1ab_0
importlib_resources=6.5.2=pyhd8ed1ab_0
intel-openmp=2022.0.1=h06a4308_3633
jinja2=3.1.6=pyhd8ed1ab_0
keyutils=1.6.1=h166bdaf_0
kiwisolver=1.4.7=py39h74842e3_0
krb5=1.21.3=h659f571_0
lame=3.100=h166bdaf_1003
lcms2=2.17=h717163a_0
ld_impl_linux-64=2.43=h712a8e2_4
lerc=4.0.0=h0aef613_1
libabseil=20250127.1=cxx17_hbbce691_0
libasprintf=0.24.1=h8e693c7_0
libblas=3.9.0=16_linux64_mkl
libbrotlicommon=1.1.0=hb9d3cd8_2
libbrotlidec=1.1.0=hb9d3cd8_2
libbrotlienc=1.1.0=hb9d3cd8_2
libcblas=3.9.0=16_linux64_mkl
libclang-cpp20.1=20.1.6=default_h1df26ce_0
libclang13=20.1.6=default_he06ed0a_0
libcublas=11.11.3.6=0
libcufft=10.9.0.58=0
libcufile=1.14.0.30=4
libcups=2.3.3=h4637d8d_4
libcurand=10.3.10.19=0
libcusolver=11.4.1.48=0
libcusparse=11.7.5.86=0
libdeflate=1.24=h86f0d12_0
libdrm=2.4.124=hb9d3cd8_0
libedit=3.1.20250104=pl5321h7949ede_0
libegl=1.7.0=ha4b6fd6_2
libexpat=2.7.0=h5888daf_0
libffi=3.4.6=h2dba641_1
libfreetype=2.13.3=ha770c72_1
libfreetype6=2.13.3=h48d6fc4_1
libgcc=15.1.0=h767d61c_2
libgcc-ng=15.1.0=h69a702a_2
libgettextpo=0.24.1=h5888daf_0
libgfortran=15.1.0=h69a702a_2
libgfortran-ng=15.1.0=h69a702a_2
libgfortran5=15.1.0=hcea5267_2
libgl=1.7.0=ha4b6fd6_2
libglib=2.84.2=h3618099_0
libglvnd=1.7.0=ha4b6fd6_2
libglx=1.7.0=ha4b6fd6_2
libgomp=15.1.0=h767d61c_2
libiconv=1.18=h4ce23a2_1
libidn2=2.3.8=ha4ef2c3_0
libjpeg-turbo=3.1.0=hb9d3cd8_0
liblapack=3.9.0=16_linux64_mkl
libllvm20=20.1.6=he9d0ab4_0
liblzma=5.8.1=hb9d3cd8_1
libnpp=11.8.0.86=0
libnsl=2.0.1=hd590300_0
libntlm=1.8=hb9d3cd8_0
libnvjitlink=12.1.105=0
libnvjpeg=11.9.0.86=0
libopengl=1.7.0=ha4b6fd6_2
libpciaccess=0.18=hd590300_0
libpng=1.6.47=h943b412_0
libpq=17.5=h27ae623_0
libsqlite=3.49.2=hee588c1_0
libstdcxx=15.1.0=h8f9b012_2
libstdcxx-ng=15.1.0=h4852527_2
libtasn1=4.20.0=hb9d3cd8_0
libtiff=4.7.0=hf01ce69_5
libunistring=0.9.10=h7f98852_0
libuuid=2.38.1=h0b41bf4_0
libva=2.22.0=h4f16b4b_2
libvpx=1.13.1=h59595ed_0
libwebp=1.5.0=hae8dbeb_0
libwebp-base=1.5.0=h851e524_0
libxcb=1.17.0=h8a09558_0
libxcrypt=4.4.36=hd590300_1
libxkbcommon=1.10.0=h65c71a3_0
libxml2=2.13.8=h4bc477f_0
libxslt=1.1.39=h76b75d6_0
libzlib=1.3.1=hb9d3cd8_2
llvm-openmp=15.0.7=h0cdce71_0
markupsafe=3.0.2=py39h9399b63_1
matplotlib=3.9.4=py39hf3d152e_0
matplotlib-base=3.9.4=py39h16632d1_0
mkl=2022.1.0=hc2b9512_224
mpc=1.3.1=h24ddda3_1
mpfr=4.2.1=h90cbb55_3
mpmath=1.3.0=pyhd8ed1ab_1
munkres=1.1.4=pyh9f0ad1d_0
natsort=8.4.0=pyh29332c3_1
ncurses=6.5=h2d0b736_3
nettle=3.9.1=h7ab15ed_0
networkx=3.2.1=pyhd8ed1ab_0
numpy=1.26.4=pypi_0
ocl-icd=2.3.3=hb9d3cd8_0
opencl-headers=2024.10.24=h5888daf_0
opencv-python=4.11.0.86=pypi_0
openh264=2.3.1=hcb278e6_2
openjpeg=2.5.3=h5fbd93e_0
openldap=2.6.10=he970967_0
openssl=3.5.1=h7b32b05_0
p11-kit=0.24.1=hc5aa10d_0
packaging=25.0=pyh29332c3_1
pcre2=10.45=hc749103_0
pillow=11.2.1=py39h15c0740_0
pip=25.1.1=pyh8b19718_0
pixman=0.46.0=h29eaf8c_0
protobuf=5.29.3=py39hbeaf701_0
pthread-stubs=0.4=hb9d3cd8_1002
pycparser=2.22=pyh29332c3_1
pyparsing=3.2.3=pyhd8ed1ab_1
pyside6=6.9.0=py39h0383914_0
pysocks=1.7.1=pyha55dd90_7
python=3.9.22=h85ef794_1_cpython
python-dateutil=2.9.0.post0=pyhff2d567_1
python_abi=3.9=7_cp39
pytorch=2.4.1=py3.9_cuda11.8_cudnn9.1.0_0
pytorch-cuda=11.8=h7e8668a_6
pytorch-msssim=1.0.0=pyhd8ed1ab_0
pytorch-mutex=1.0=cuda
pyyaml=6.0.2=py39h9399b63_2
qhull=2020.2=h434a139_5
qt6-main=6.9.0=h0384650_3
readline=8.2=h8c095d6_2
requests=2.32.3=pyhd8ed1ab_1
safetensors=0.5.3=py39he612d8f_0
scipy=1.13.1=py39haf93ffa_0
setuptools=80.1.0=pyhff2d567_0
six=1.17.0=pyhd8ed1ab_0
svt-av1=1.4.1=hcb278e6_0
sympy=1.13.1=pypi_0
tensorboardx=2.6.2.2=pyhd8ed1ab_1
thop=0.1.1-2209072238=pypi_0
timm=1.0.15=pyhd8ed1ab_0
tk=8.6.13=noxft_h4845f30_101
torchaudio=2.4.1=py39_cu118
torchtriton=3.0.0=py39
torchvision=0.19.1=py39_cu118
tornado=6.5.1=py39h8cd3c5a_0
tqdm=4.67.1=pyhd8ed1ab_1
typing-extensions=4.13.2=h0e9735f_0
typing_extensions=4.13.2=pyh29332c3_0
tzdata=2025b=h78e105d_0
unicodedata2=16.0.0=py39h8cd3c5a_0
urllib3=2.4.0=pyhd8ed1ab_0
wayland=1.23.1=h3e06ad9_1
wayland-protocols=1.43=hd8ed1ab_0
wheel=0.45.1=pyhd8ed1ab_1
x264=1!164.3095=h166bdaf_2
x265=3.5=h924138e_3
xcb-util=0.4.1=hb711507_2
xcb-util-cursor=0.1.5=hb9d3cd8_0
xcb-util-image=0.4.0=hb711507_2
xcb-util-keysyms=0.4.1=hb711507_0
xcb-util-renderutil=0.3.10=hb711507_0
xcb-util-wm=0.4.2=hb711507_0
xkeyboard-config=2.44=hb9d3cd8_0
xorg-libice=1.1.2=hb9d3cd8_0
xorg-libsm=1.2.6=he73a12e_0
xorg-libx11=1.8.12=h4f16b4b_0
xorg-libxau=1.0.12=hb9d3cd8_0
xorg-libxcomposite=0.4.6=hb9d3cd8_2
xorg-libxcursor=1.2.3=hb9d3cd8_0
xorg-libxdamage=1.1.6=hb9d3cd8_0
xorg-libxdmcp=1.1.5=hb9d3cd8_0
xorg-libxext=1.3.6=hb9d3cd8_0
xorg-libxfixes=6.0.1=hb9d3cd8_0
xorg-libxi=1.8.2=hb9d3cd8_0
xorg-libxrandr=1.5.4=hb9d3cd8_0
xorg-libxrender=0.9.12=hb9d3cd8_0
xorg-libxtst=1.2.5=hb9d3cd8_3
xorg-libxxf86vm=1.1.6=hb9d3cd8_0
yaml=0.2.5=h7f98852_2
zipp=3.22.0=pyhd8ed1ab_0
zstandard=0.23.0=py39h8cd3c5a_2
zstd=1.5.7=hb8e6e7a_2
